backend/requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
spotipy
google-genai
